SessionBus if DBUS_SESSION_BUS_ADDRESS in env, else System
Previously, the code looked at the machine type (armv7a). Which works
fine on a CCGX, but not on any other type, for example a bbb. New way
works better, and is also the same mechanism used by C code.

The purpose of the autoselection between Session and System is easy
debugging: on a desktop use the Sessionbus, and on targets use System.