prefix xml elements with an underscore when starting with a digit
XML elements cannot start with a digit, and therefore localsettings
did not support settings of which one of the sections in the path
started with a number. For example adding /Settings/Tank/0/Type
caused localsettings to show an error while saving the file:

Traceback (most recent call last):
  File "./localsettings.py", line 401, in saveSettingsCallback
    parseDictonaryToXmlFile(settings, fileSettings)
  File "./localsettings.py", line 488, in parseDictonaryToXmlFile
    elem = etree.SubElement(elem, item)
  File "lxml.etree.pyx", line 2935, in lxml.etree.SubElement (src/lxml/lxml.etree.c:67055)
  File "apihelpers.pxi", line 181, in lxml.etree._makeSubElement (src/lxml/lxml.etree.c:15299)
  File "apihelpers.pxi", line 1575, in lxml.etree._tagValidOrRaise (src/lxml/lxml.etree.c:28606)
ValueError: Invalid tag name u'2'

Now, when saving to the xml file on disk, all items that start with
a digit are prefixed with an underscore. And, when loading from disk,
that underscore is removed again.

This change does not cause any migration issues, since there were,
obviously, no settings yet that started with a digit.