AddSetting: fix error when re-adding a setting with different params
Adding, perhaps updating is a better word, a setting that already exists resulted
in an error because myDbusObject wasn't assigned in that case. The error only
occured when re-adding a setting with a different min, max and/or default than
used in the original creation.

This bug was introduced in v1.02, released in November 2015.

Also remove some unnecessary nesting as well as a try/catch that doesn't make any
sense.