rename DeviceInstance to VrmDeviceInstance
Since localsettings can now assign instances to services, there are
in the CAN-bus world now two different device instances; the VE.Can
device instance and the instance assigned to a device by localsettings.
The later is now called VRM device instance, to discriminate between
the two, so rename the method to make it explicit which instance is
dealt with.