add the group hierarchy
Groups have a hierarchy, both on the dbus and in the xml, so lets add that
as a start, so we can eventually stop scanning the global settings all the
time. While at it, make the root object a group object. That also allows
calling AddSetting on the root. The vedbus doesn't know about hierarchy
either, so remove the suggestion to use it. Add members to allocate new
Groups and Settings so derived class an override it.