add RemoveSettings
Add the RemoveSettings dbus call which removes a setting and the
parent groups if they become empty. Never delete the root nor /Settings.

The call is usefull for testing if Settings are propertly freed.
Since the dbus.service.Object constructor adds a reference to the
object, it must be explicitly removed again or it will never be freed
and no new object can be added to the path.

It can be called with e.g.

dbus com.victronenergy.settings / RemoveSettings '%["/Settings/Gui/Language"]'