allow _setValue to fail
The upcoming device mapper doesn't allow arbitrary values to set. To do
that addSetting and SetValue would need to be overriden at the moment
and perform the same checks. Since both call _setValue allow that
function to fail instead.

If a new settingObject was allocated, make sure it is properly removed,
since the dbus connection has a reference to it. If it was an existing
setting, don't remove it, just report the error.