adjust cgwacs migration
The migration for cgwacs was included in a candadite while the actual
application change was not. Hence the old settings are recreated and
both the old and new settings are present. If the migration is done
again, the newly created settings will overwrite the valid one. Prevent
that by not updating the new settings if they are already present.
While at it, prefix the device numbers since it was missing.