emit a signal if a setting is removed
While there is a standard signal for notifying that a object path is removed,
it is not listened to. When a setting gets removed, at least send a signal that
the setting is no longer valid.

This is mainly for testing purposes, there are no uses cases at the moment for
removing settings.