python3 64bit awareness
In python3, there is no longer a longint type, which causes python-dbus
to assume anything we are not explicit about is an int32. This causes
problems with large numbers.

We now have to to be explicit about the types we return. So properly
wrap returned vales as dbus types.

The wrapper in velib_python is not used so as to not introduce a
dependency that wasn't there before. Presently, velib_python is
included only to facilitate testing.