Fix the signature of dbus GetText call on /
The velib C and C++ implementation of the dbus producers return a
variant containing a dict(string, string) when calling GetText on the
root (or any other non-leaf item). This is also what the C++ consumer
(like in the gui) expects and what is needed for the bulk text init to
work. By doing this also in localsettings, the bulk text init also works
for localsettings and some GetText calls can be saved when the gui
starts.