support class changes in the default value of a ClassAndVrmInstance
velib offers a function to fetch the device instance from localsettings.
This always uses AddSettings and the default contains the requested
class and instance. It was however never properly supported that this
would change, for example when the monitor mode of a SmartShunt is
changed. Localsettings would only update the default, but return the
existing value, which would contain the wrong class. As the velib code
only checked the instance and not the class that was returned, this
could cause multiple services to use the same instance.

With this commit, localsettings checks whether the class of the provided
default is different from the current value and if so, changes the value
to the new class and provides a proper unique instance for the new
class.