support forceValue argument on AddSettings
When this is true'ish, it forces the value to become the same as
default, even when a value has already been set.

For ClassAndVrmInstance settings, the instance will be adjusted such
that it returns a free instance, but only when the class of the value
actually changes. When the value already contained the new class, the
instance value is kept.