allow ClassAndVrmInstance to have no default
Old configuration files often have ClassAndVrmInstance tags
with no default. This causes a NoneType exception when such a file
is loaded in Venus 3.20 or later. Since this is only part of a check
to ensure that something illegal is not set, simply skip the check
if it is not specified.

This specifically affects migration of cgwacs meters from before
Venus 2.60 to Venus 3.20 or later.

https://github.com/victronenergy/venus/issues/1332