fix energy meters and pvinverters being lost
A migration added in Venus 2.60 migrated the DeviceInstance of
CGwacs and Fronius devices to the new ClassAndVrmInstance scheme.
This did not add a default attribute to the tag. This is still not
done in Venus 3.50. A change in Venus 3.20 caused such settings
(created before Venus 2.60) to be corrupted. This causes the
devices not to show up at all, with no way to fix it.

This was partially fixed in 85acef. It no longer corrupts the
settings. But settings files that are already corrupted remain
broken.

To fix this, simply delete empty ClassAndVrmInstance tags. It
loses no information that has not already been lost.

https://github.com/victronenergy/venus-private/issues/282