allow empty strings in a settings file
Since the line is split at the moment on any whitespace an empty string
can't be used as a default. So lets encourage quoting strings. As is
whitespace inside the string is not supported for the same reason.
That can be added, but would require changing how the line is parsed.