do not reset the saving timer
localsettings used to wait till no settings are changed for a while,
before saving the settings. Since nodered can now be used to change
IO pins e.g. there might be a continues flow of setting changes
causing the settings file to be never saved. Not resetting the timer
does prevent that.