repeat migration of dess limits
In a story for the ages, we released localsettings v1.64 in Venus
v3.60~33. This would migrate the settings to version 16, where
the dess limits are floats.

However, the systemcalc version that would create them as float,
was only released in v3.60~47. So between beta 33 and 47, a lot of
people missed the type change, as the settings version was already
at 16 or above.

Since these are old betas, there is no real reason anyone would
want to downgrade to those, and doing so does not cause harm.
So it is alright for this migration to be somewhat blunt and
simply change the type in place. Changing the paths is not
really an option as it will require deep changes in the DESS
implementation on VRM as well.